Move data transformation from notebook to action
This commit moves data transformation code from the notebook
to the concat_study_populations action, making it easier to maintain and
(eventually) to test. Mostly, the move is a copy-paste, although some
comments for a non-technical reader have been removed and data frame
column access has been standardised, with, for example `records["age"]`
preferred to `records.age`.

This commit, however, introduces a regression: age groups are now
represented by a zero-based integer index rather than a category,
because of a "quirk" in how Pandas writes and reads feather files.
Whilst Pandas is happy to write a feather file with a
categorical column that's the result of calling `pandas.cut`, it isn't
happy to read the feather file, complaining of duplicate categories.
As the shape of the distribution is more important than the category
labels, this regression seems reasonable.